requirements.txt
aiohttp==3.11.10
asyncpg==0.30.0
Babel==2.8.0
dateparser==1.2.0
discord.py==2.4.0
distro==1.7.0
fantasy_names==1.0.0
firebase_admin==6.6.0
humanize==4.12.3
lru_dict==1.3.0
matplotlib==3.10.5
numpy==2.3.2
openai==1.98.0
orjson==3.11.1
Pillow==11.3.0
psutil==6.1.0
pytesseract==0.3.13
pytz==2024.2
redis==5.2.1
Requests==2.32.4
seaborn==0.13.2
sentry_sdk==2.19.2
setuptools==59.6.0
sly==0.5
textblob==0.19.0
tomli==2.2.1
uvloop==0.21.0
validators==0.34.0
yarl==1.18.3
